openclaw-knowledge-radio/requirements.txt
PyYAML==6.0.2
requests==2.32.3
feedparser==6.0.11
python-dateutil==2.9.0.post0
edge-tts==6.1.12
openai==1.40.0
